lib: fix determining our own realm
krb5_sname_to_principal() may legitimately return an empty realm if our
domain is not listed in the [domain_realm] section. (During actual
authentication, Krb5 would either check _kerberos via DNS or just send
the null realm to a KDC and receive a referral.)

However, context->sname is not in fact used anywhere else, and I think
it's safe to assume that in most cases the host's realm will be the same
as the system default realm, so let's just get rid of it.

(I'll try to make the realm configurable for the case where it isn't.)

Ref: #6